(client, migrations, routes) Alter score column to BIGINT, Make sure form resets in modal when reopening

+++ routes/scores/get-scores-by-game-id.ts
@@ -23,7 +23,7 @@ const getScoresByGameId = async (pool: DatabasePoolType, gameId: string): Promis
         WHERE
             s.game_id = ${gameId} AND
             s.is_deleted = FALSE
-        ORDER BY s.score::INT DESC
+        ORDER BY s.score DESC
         LIMIT 5;
     `)
